# Improve formatting performance