--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/InterlockedOperations.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/InterlockedOperations.cs
@@ -1,0 +1,15 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Diagnostics.CodeAnalysis;
+using System.Threading;
+
+namespace Microsoft.AspNetCore.Razor;
+
+internal static class InterlockedOperations
+{
+    public static T Initialize<T>([NotNull] ref T? target, T value) where T : class
+    {
+        return Interlocked.CompareExchange(ref target, value, null) ?? value;
+    }
+}