… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/Syntax/MarkupStartTagSyntax.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/Syntax/MarkupTagHelperEndTagSyntax.cs
… src/Compiler/Microsoft.AspNetCore.Razor.Language/src/Syntax/MarkupTagHelperStartTagSyntax.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/AutoInsert/AutoClosingTagOnAutoInsertProvider.cs
… src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/LinkedEditingRange/LinkedEditingRangeEndpoint.cs
… src/Razor/src/Microsoft.VisualStudio.Editor.Razor/BraceSmartIndenter.cs